move error handling to middleware
