+++ config/templates/default.yaml
@@ -43,3 +43,4 @@ rag:
 datasets: []
 models: []
 
+
