--- server/api/models.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-
-
-class ErrorResponse(BaseModel):
-    error: str
-    message: str
-    details: str | None = None
-
-
